feat(links): Created Icon Link Element
Tested an idea for creating links to external pages with the fav icon of that page.
It works, however the styling is weird as I'd like a black and white icon not the colored favicon.
This is not being added to the current landing page but it's a fun idea in another style